requirement.txt
streamlit
tensorflow
Pillow
numpy==2.2.6
pandas==2.3.1
pip==22.3.1
gtts
gdown